% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.SVCMimage.R
\name{plot.SVCMimage}
\alias{plot.SVCMimage}
\title{Produces coefficient function plots for a fitted "SVCMimage" object.}
\usage{
\method{plot}{SVCMimage}(mfit)
}
\arguments{
\item{mfit}{Fitted ``SVCMimage" object.
\cr}
}
\value{
None
}
\description{
This function produces the images of the estimated coefficient function for multivariate varying coefficient models.
}
