% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.SVCMimage.R
\name{predict.SVCMimage}
\alias{predict.SVCMimage}
\title{Make predictions from a fitted multivariate varying coefficient models.}
\usage{
\method{predict}{SVCMimage}(mfit, Xpred, Zpred)
}
\arguments{
\item{mfit}{Fitted ``SVCMimage" object.
\cr}

\item{Xpred}{The design matrix for prediction.
\cr}

\item{Zpred}{The cooridinates for prediction.
\cr}
}
\value{
A matrix of predicted images is returned.
}
\description{
This function is used to make predictions of the multivariate varying coefficient models.
}
\details{
This R package is the implementation program for manuscript entitled ``Multivariate Spline Estimation and Inference for Varying Coefficient Models with Imaging Data" by Shan Yu, Guannan Wang, Li Wang and Lijian Yang.
}
